Reference

Public API

The following functions are exported by the JefimenkoModels package.

JefimenkoModels.E — Function

H(r̄::AbstractCoordinate, t::Time, model::JefimenkoModel; rtol=sqrt(eps))

Calculate the predicted electric field 𝐇 observed at space-time point (r̄,t) using the electric Jefimenko equation for a particular model. Calculate the integral using a specified relative tolerance.

Arguments

r̄::UnitfulCoordinateSystems.AbstractCoordinate: spatial location of the observation point

t::Unitful.Time: time at which the electric field is observed

model::JefimenkoModel: model of the transmitting source and propagation media

Keywords

rtol::Real: relative tolerance at which to solve the integral (optional)

JefimenkoModels.H — Function

H(r̄::AbstractCoordinate, t::Time, model::JefimenkoModel; rtol=sqrt(eps))

Calculate the predicted magnetic field 𝐇 observed at space-time point (r̄,t) using the magnetic Jefimenko equation for a particular model. Calculate the integral using a specified relative tolerance.

Arguments

r̄::UnitfulCoordinateSystems.AbstractCoordinate: spatial location of the observation point

t::Unitful.Time: time at which the field is observed

model::JefimenkoModel: model of the transmitting source and propagation media

Keywords

rtol::Real: relative tolerance at which to solve the integral (optional)

JefimenkoModels.P — Function

P(r̄::AbstractCoordinate, t::Time, model::JefimenkoModel; rtol=sqrt(eps))

Calculate the predicted Poynting vector 𝐏 observed at space-time point (r̄,t) using the electric and magnetic Jefimenko equations for a particular model. Calculate the integrals using a specified relative tolerance.

Arguments

r̄::UnitfulCoordinateSystems.AbstractCoordinate: spatial location of the observation point

t::Unitful.Time: time at which the field is observed

model::JefimenkoModel: model of the transmitting source and propagation media

Keywords

rtol::Real: relative tolerance at which to solve the integral (optional)

Internal API

JefimenkoModels.__E — Function

__E(r̄::AbstractCoordinate, t::Time, source::AbstractJefimenkoSource,
    media::PropagationMedia; rtol=sqrt(eps))

Calculate the electric field at (r̄,t) using the electric Jefimenko equation due to a particular source, transmitted through a particular homogeneous propagation media. Calculate the integral using a specified relative tolerance.

Arguments

r̄::UnitfulCoordinateSystems.AbstractCoordinate: spatial location of the observation point

t::Unitful.Time: time at which the electric field is observed

source::AbstractJefimenkoSource{T}: source of the electric field

media::PropagationMedia: properties of the propagation media

Keywords

rtol::Real: relative tolerance at which to solve the integral (optional)

JefimenkoModels.__integrand_E — Function

__integrand_E(r̄′::CoordinateCartesian; r̄::CoordinateCartesian, t::Time,
              source::AbstractJefimenkoSource{T}, media::PropagationMedia)::SVector{3,T}
                    where {T<:AbstractFloat}

Calculate the integrand function for the electric Jefimenko equation at the source point r̄′. Parameterize the integrand function according to a particular field source, propagation media, and for an observer positioned at space-time point (r̄,t).

Arguments

r̄′::UnitfulCoordinateSystems.CoordinateCartesian: coordinate of the source point

Parameters

r̄::UnitfulCoordinateSystems.CoordinateCartesian: coordinate of the observation point

t::Unitful.Time: time at the observation point

source::JefimenkoSource: the source model generating the electric field

media::PropagationMedia_Simple: properties of the propagation media

Returns

SVector{3,T}: the predicted vector-valued integrand value

JefimenkoModels.__H — Function

__H(r̄::AbstractCoordinate, t::Time, source::JefimenkoSource, media::PropagationMedia; rtol)

Calculate the magnetic field at (r̄,t) using the electric Jefimenko equation due to a particular source, transmitted through a particular homogeneous propagation media. Calculate the integral using a specified relative tolerance.

Arguments

r̄::UnitfulCoordinateSystems.AbstractCoordinate: spatial location of the observation point

t::Unitful.Time: time at which the magnetic field is observed

source::JefimenkoSource: source of the magnetic field

media::PropagationMedia: properties of the propagation media

Keywords

rtol::Real: relative tolerance at which to solve the integral (optional)

JefimenkoModels.__integrand_H — Function

__integrand_H(r̄′::CoordinateCartesian; r̄::CoordinateCartesian, t::Time,
              source::AbstractJefimenkoSource{T}, media::PropagationMedia)::SVector{3,T}
                    where {T<:AbstractFloat}

Calculate the integrand function for the magnetic Jefimenko equation at the source point r̄′. Parameterize the integrand function according to a particular field source, propagation media, and for an observer positioned at space-time point (r̄,t).

Arguments

r̄′::UnitfulCoordinateSystems.CoordinateCartesian: coordinate of the source point

Parameters

r̄::UnitfulCoordinateSystems.CoordinateCartesian: coordinate of the observation point

t::Unitful.Time: time at the observation point

source::JefimenkoSource: the source model generating the magnetic field

media::PropagationMedia_Simple: properties of the propagation media

Returns

SVector{3,T}: the predicted vector-valued integrand value

JefimenkoModels.__P — Function

__P(r̄::AbstractCoordinate, t::Time, source::JefimenkoSource, media::PropagationMedia; rtol)

Calculate the predicted Poynting vector 𝐏 observed at space-time point (r̄,t) due to a particular source, transmitted through a particular propagation media. Calculate the integral using a specified relative tolerance.

Arguments

r̄::UnitfulCoordinateSystems.AbstractCoordinate: spatial location of the observation point

t::Unitful.Time: time at which the electric field is observed

source::JefimenkoSource: source of the electric field

media::PropagationMedia: properties of the propagation media

Keywords

rtol::Real: relative tolerance at which to solve the integral (optional)